+++ packages/wallet/src/types.ts
@@ -1,5 +1,3 @@
-import { KeyringControllerOptions } from '@metamask/keyring-controller';
-import { StorageAdapter } from '@metamask/storage-service';
 import type { Json } from '@metamask/utils';
 
 import type {
@@ -8,7 +6,8 @@ import type {
   RootMessenger,
 } from './initialization/defaults';
 import type { ApprovalControllerInstanceOptions } from './initialization/instances/approval-controller/types';
-import { GenericEncryptor } from './initialization/instances/keyring-controller';
+import type { KeyringControllerInstanceOptions } from './initialization/instances/keyring-controller/types';
+import type { StorageServiceInstanceOptions } from './initialization/instances/storage-service/types';
 import { InitializationConfiguration } from './initialization/types';
 
 export type WalletOptions = {
@@ -23,12 +22,6 @@ export type WalletOptions = {
 
 export type InstanceSpecificOptions = {
   approvalController?: ApprovalControllerInstanceOptions;
-  keyringController?: {
-    encryptor?: GenericEncryptor;
-    keyringBuilders?: KeyringControllerOptions['keyringBuilders'];
-    keyringV2Builders?: KeyringControllerOptions['keyringV2Builders'];
-  };
-  storageService: {
-    storage: StorageAdapter;
-  };
+  keyringController?: KeyringControllerInstanceOptions;
+  storageService: StorageServiceInstanceOptions;
 };